+++ CHANGES.md
@@ -5,7 +5,7 @@
 
 ### Version 8.13.5, Not yet released
 
-  * Current development version
+  * Improve the all-fast-sat algorithm to also support uninterpreted values.
 
 ### Version 8.13, 2021-03-21